arda's code, had to commit this because porting it back from a non-maven project is too difficult

+++ src/main/java/domain/entities/Knight.java
@@ -37,7 +37,7 @@ public class Knight extends Enemy {
 	}
 
 	@Override
-	public void updateEnemy(long deltaTime) {
+	public void updateEnemy(double deltaTime) {
 		super.updateEnemy(deltaTime);
 		knightSpeedUp();
 	}
